refactor: anti-barrel
