--- packages/salesforcedx-vscode-core/src/salesforceProject/index.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2019, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-export { default as SalesforcePackageDirectories } from './salesforcePackageDirectories';
-export { default as SalesforceProjectConfig } from './salesforceProjectConfig';
